version 2

+++ ex_03.py
@@ -1,8 +1,8 @@
 #Write a program to prompt the user for hours and rate per hour to compute
 #gross pay.
-hours = int(input("Enter Hours: "))
+hours = float(input("Enter Hours: "))
 rate = float(input("Enter Rate: "))
 
 pay = hours * rate
- 
+
 print("pay: ", pay)
